refactor: rename `pyjarchery` package to `pyarchery` and update related files
